added a check for Enable_OTel Enviromental variable
This adds a check for a enviromental variable to explicitly enable OTEL, other wise set the application to output to stdout. in conjunction with ENABLE_OTEL you have to have the OTEL as a default feature,
if you don't have Otel as feature need to add tracing-subscriber as one.
this could probably be better